% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{cross_validated_mse_for_subset_selection}
\alias{cross_validated_mse_for_subset_selection}
\title{Perform subset selection using cross-validated errors to evaluate the best model}
\usage{
cross_validated_mse_for_subset_selection(
  data,
  response,
  K = 10,
  method = "forward",
  nvmax = NULL,
  plot_mean_mse = T,
  ncores = 1
)
}
\arguments{
\item{data}{Matrix or data.frame containing the response and features as columns.}

\item{response}{Name of the column containing the response in data.}

\item{K}{Number of folds. Default is 10.}

\item{method}{Type of feature selection to perform. Same as for leaps::regsubsets. Default is forward selection.}

\item{nvmax}{Maximum number of features to consider. Default is all features in data.}

\item{plot_mean_mse}{A logical value indicating whether to plot the mean mse for each size model. Default is TRUE.}

\item{ncores}{Number of cores to use to process folds separately in parallel. Default is process each fold sequentially.
Cannot be greater than K.}
}
\value{
A data.frame with the CV MSE for best model of the indicated size found for each fold of data.
}
\description{
Perform subset selection using cross-validated errors to evaluate the best model
}
