% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_regression.R
\name{extract_best_features_from_lasso}
\alias{extract_best_features_from_lasso}
\title{Extract the names of the features selected by lasso regression}
\usage{
extract_best_features_from_lasso(lasso_model)
}
\arguments{
\item{lasso_model}{An object of class glmnet produced by fit_best_lasso_with_cv()}
}
\value{
A character vector with the names of the selected features.
}
\description{
Extract the names of the features selected by lasso regression
}
