% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{extract_best_model_from_subset_selection}
\alias{extract_best_model_from_subset_selection}
\title{Create a lm object from the best identified model of a specified size by subset selection}
\usage{
extract_best_model_from_subset_selection(regsubsets, n, data)
}
\arguments{
\item{regsubsets}{A regsubsets object.}

\item{n}{The size of the model.}

\item{data}{Data used to fit the model. Should be the same as used to create the regsubsets object.}
}
\value{
An lm object.
}
\description{
Create a lm object from the best identified model of a specified size by subset selection
}
