% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_regression.R
\name{fit_best_lasso_with_cv}
\alias{fit_best_lasso_with_cv}
\title{Fit a lasso regression model using cross-validation to identify the best value for the tuning parameter.}
\usage{
fit_best_lasso_with_cv(data, response, lambda = NULL, K = 10, ncores = 1)
}
\arguments{
\item{data}{Matrix or data.frame containing the response and features as columns.}

\item{response}{Name of the column containing the response in data.}

\item{lambda}{A vector of values for the tuning parameter to be tested using cross-validation.
Default is to let glmnet choose its own values.}

\item{K}{Number of folds. Default is 10.}

\item{ncores}{Number of cores to use to process folds separately in parallel. Default is process each fold sequentially.
Cannot be greater than K.}
}
\value{
An object of class glmnet for lasso regression.
}
\description{
Fit a lasso regression model using cross-validation to identify the best value for the tuning parameter.
}
