% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_methods.R
\name{remove_constant_features}
\alias{remove_constant_features}
\title{Remove constant features from a dataset}
\usage{
remove_constant_features(data)
}
\arguments{
\item{data}{Any tabular data structure.}
}
\value{
The dataset with any constant features removed.
}
\description{
Remove constant features from a dataset
}
