% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{rf_var_importance}
\alias{rf_var_importance}
\title{Calculate variable importance using a random forest.}
\usage{
rf_var_importance(data, response, mtry = NULL, ntree = 500)
}
\arguments{
\item{data}{Matrix or data.frame containing the response and features as columns.}

\item{response}{Name of the column containing the response in data.}

\item{mtry}{Number of features used at each split. Uses same default as randomForest::randomForest().}

\item{ntree}{Number of trees. Default is 500.}
}
\value{
A matrix with the feature importance measures.
}
\description{
Calculate variable importance using a random forest.
}
