% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{sample_k_fold_indices}
\alias{sample_k_fold_indices}
\title{Create row indices for K folds of a dataset}
\usage{
sample_k_fold_indices(data, K = 10)
}
\arguments{
\item{data}{Dataset to split into folds}

\item{K}{Number of folds. Default is 10.}
}
\value{
A list with the row indices for K folds of data
}
\description{
Create row indices for K folds of a dataset
}
