% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{select_best_model_size_from_cv_subset_results}
\alias{select_best_model_size_from_cv_subset_results}
\title{Select the best model size from cross-validated errors for subset selection}
\usage{
select_best_model_size_from_cv_subset_results(
  cv_subset_results,
  one_sd_error_rule = FALSE
)
}
\arguments{
\item{cv_subset_results}{A data.frame with the results of cross_validated_mse_for_subset_selection.}

\item{one_sd_error_rule}{A logical value indicating whether to use the one-standard-error rule to select the smallest
model where the mean CV MSE is within one standard error of the model with the lowest mean CV MSE. Default is FALSE.}
}
\value{
An integer with the identified best model size.
}
\description{
Select the best model size from cross-validated errors for subset selection
}
