% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{separate_training_and_test_data}
\alias{separate_training_and_test_data}
\title{Separate a dataset into a training and test set using specified folds of the data}
\usage{
separate_training_and_test_data(data, k_folds_indices, test_fold)
}
\arguments{
\item{data}{Matrix or table to split into folds}

\item{k_folds_indices}{A list with the indices for different folds of the data.
Each index should occur in only one fold.}

\item{test_fold}{Number of the fold from k_folds_indices which should be treated as the test set.}
}
\value{
A list with the training and test set splits of the data.
}
\description{
Separate a dataset into a training and test set using specified folds of the data
}
